+++ webdriver-precore/test/BiDi/DemoUtils.hs
@@ -1,6 +1,6 @@
 module BiDi.DemoUtils where
 
-import BiDi.BiDiActions (BiDiActions (..), mkActions)
+import BiDi.Actions (BiDiActions (..), mkActions)
 import BiDi.Runner (withBiDi, withBidiFailTest)
 import BiDi.Socket (SocketActions)
 import Config (Config (..))
@@ -12,7 +12,7 @@ import Data.Text (Text, isInfixOf, unpack)
 import Data.Time.Clock.POSIX (POSIXTime, getPOSIXTime)
 import Data.Word (Word64)
 import HTTP.DemoUtils (withSession)
-import HTTP.HttpActions qualified as HTTPA
+import HTTP.Actions qualified as HTTPA
 import HTTP.Runner (mkRunner)
 import IOUtils (DemoActions (..), Logger, logNothingLogger, mkDemoActions)
 import Logger (withChannelFileLogger)