+++ webdriver-precore/test/BiDi/Demos/NetworkEventDemos.hs
@@ -1,6 +1,6 @@
 module BiDi.Demos.NetworkEventDemos where
 
-import BiDi.BiDiActions (BiDiActions (..))
+import BiDi.Actions (BiDiActions (..))
 import BiDi.DemoUtils
 import IOUtils (DemoActions (..))
 import TestServerAPI (authTestUrl, malformedResponseUrl, testServerHomeUrl, withTestServer)