+++ webdriver-precore/test/HTTP/DemoUtils.hs
@@ -7,7 +7,7 @@ import ConfigLoader (loadConfig)
 import Const (milliseconds)
 import Control.Exception (bracket)
 import Data.Text (Text)
-import HTTP.HttpActions (HttpActions (..))
+import HTTP.Actions (HttpActions (..))
 import IOUtils (DemoActions (..), Logger, logNothingLogger, mkDemoActions)
 import Logger (withChannelFileLogger)
 import Network.HTTP.Req (http)
@@ -19,7 +19,7 @@ import HTTP.HttpActionsDeprecated qualified as Legacy
 import HTTP.HttpRunnerDeprecated qualified as Legacy
 #else
 import HTTP.Runner (mkRunner)
-import HTTP.HttpActions (mkActions)
+import HTTP.Actions (mkActions)
 #endif
 
 data HttpDemo